Firefox support (#5)
* feat: add firefox support with browser-polyfill

* fix: set explicit id in manifest for firefox usage

* fix: chrome compatibility when using the webextension-polyfill

* fix: replace callbacks with promise chains in all webextension-polyfill function calls

Support chrome while using webextension-polyfill

* revert: {add background.persistent: false} back into manifest

After reviewing documentation, this feature may not work in older versions of firefox, but does work in chrome. There is not downside to keeping it in the manifest despite warning in firefox.

* fix: autoload support in firefox

Support autoload in Firefox on the primary window on inital load, matching the current autoload support in Chrome.